Meow meow hi, I kill my wife yo, shoot my head

+++ Assets/Scripts/Interfaces - Managers/IManager.cs
@@ -20,5 +20,5 @@ public interface IManager {
 	
 	bool CostAcceptable(float cost);
 	
-	int Resource { get; }
+	float Resource { get; }
 }
